requirements.txt
Django
djangorestframework
netifaces
psycopg2